EMP-553: refactored CRM5 and CMR7 into their own model packege
