+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/crm7/Crm7DisbursementsModel.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.equinity.historicaldata.model;
+package uk.gov.justice.laa.crime.equinity.historicaldata.model.crm7;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
@@ -8,8 +8,7 @@ import java.util.List;
 
 @Data
 @NoArgsConstructor
-public class Crm7ScheduleTimeSpentModel {
-    // Schedule
+public class Crm7DisbursementsModel {
     @JsonProperty("row")
-    List<Crm7TimeSpentModel> timeSpent;
+    public List<Crm7DisbursementModel> table;
 }
